fix for error in older version of sqlalchemy migrate
found on metabuildbot:

  File "/buildbot/sqla_0_8_7-sqlam_0_9_1/sandbox-python2.7-twisted-sqlalchemy==0.8.7samigr-0.9.1/local/lib/python2.7/site-packages/migrate/changeset/databases/sqlite.py", line 185, in visit_migrate_foreign_key_constraint
    self._not_supported('ALTER TABLE DROP CONSTRAINT')
  File "/buildbot/sqla_0_8_7-sqlam_0_9_1/sandbox-python2.7-twisted-sqlalchemy==0.8.7samigr-0.9.1/local/lib/python2.7/site-packages/migrate/changeset/databases/sqlite.py", line 27, in _not_supported
    "%s; see http://www.sqlite.org/lang_altertable.html" % op)
migrate.exceptions.NotSupportedError: SQLite does not support ALTER TABLE DROP CONSTRAINT; see http://www.sqlite.org/lang_altertable.html

+++ master/buildbot/db/migrate/versions/047_cascading_deletes.py
@@ -15,6 +15,7 @@
 
 import sqlalchemy as sa
 from migrate.changeset.constraint import ForeignKeyConstraint
+
 from buildbot.util import sautils
 
 
@@ -38,5 +39,6 @@ def upgrade(migrate_engine):
                ForeignKeyConstraint([configured_workers.c.workerid],
                                     [workers.c.id], ondelete='CASCADE'),
                ):
-        fk.drop()
+        if migrate_engine.dialect.name != 'sqlite':
+            fk.drop()
         fk.create()
